Finalizado teste descritivo do quartz , com agendamento de serviço helloworld

+++ MicroServices/Integrations/QuartzScheduler/Jobs/HelloWorldJob.cs
@@ -1,7 +1,9 @@
 ﻿using Quartz;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
+using System.Reflection.Emit;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,10 +11,18 @@ namespace QuartzScheduler.Jobs
 {
     public class HelloWorldJob : IJob
     {
+        public string path = Path.Combine(Directory.GetParent(AppContext.BaseDirectory).Parent.Parent.Parent.Parent.FullName, "Task Scheduler Test.file");
         public async Task Execute(IJobExecutionContext context)
         {
-            string path =Path.Combine(Directory.GetParent(AppContext.BaseDirectory).Parent.Parent.Parent.Parent.FullName, "Task Scheduler Test");
-            File.Create(path);
+            if(!File.Exists(path))
+            {
+                File.Create(path);
+                Console.WriteLine("CREATED FILE");
+                Debug.WriteLine("CREATED FILE");
+            }
+            Console.WriteLine("FILE ALREADY CREATED");
+            Debug.WriteLine("FILE ALREADY CREATED");
+
         }
     }
 }
